Make of-builder use internal trust secret
- this prevents of-builder from being called by other services
which may be able to reach the TCP socket within the network

+++ of-builder/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"flag"
@@ -13,12 +14,14 @@ import (
 	"sync"
 	"time"
 
+	"github.com/alexellis/hmac"
 	"github.com/docker/docker/pkg/archive"
 	"github.com/gorilla/mux"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/session/auth/authprovider"
 	"github.com/moby/buildkit/util/appcontext"
+	"github.com/openfaas/openfaas-cloud/sdk"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
 )
@@ -67,7 +70,9 @@ func main() {
 }
 
 func buildHandler(w http.ResponseWriter, r *http.Request) {
+
 	dt, err := build(w, r)
+
 	if err != nil {
 		w.WriteHeader(500)
 
@@ -90,16 +95,35 @@ func buildHandler(w http.ResponseWriter, r *http.Request) {
 
 func build(w http.ResponseWriter, r *http.Request) ([]byte, error) {
 
+	if r.Body == nil {
+		return nil, fmt.Errorf("a body is required to build a function")
+	}
+
+	defer r.Body.Close()
+
 	tmpdir, err := ioutil.TempDir("", "buildctx")
 	if err != nil {
 		return nil, err
 	}
+
+	tarBytes, bodyErr := ioutil.ReadAll(r.Body)
+	if bodyErr != nil {
+		return nil, bodyErr
+	}
+
+	hmacErr := validateRequest(&tarBytes, r)
+
+	if hmacErr != nil {
+		return nil, hmacErr
+	}
+
 	defer os.RemoveAll(tmpdir)
+
 	opts := archive.TarOptions{
 		NoLchown: !lchownEnabled,
 	}
 
-	if err := archive.Untar(r.Body, tmpdir, &opts); err != nil {
+	if err := archive.Untar(bytes.NewReader(tarBytes), tmpdir, &opts); err != nil {
 		return nil, err
 	}
 
@@ -249,4 +273,22 @@ func (b *buildLog) Append(msg string) {
 
 	b.Line = append(b.Line, msg)
 
+}
+
+func validateRequest(req *[]byte, r *http.Request) (err error) {
+	payloadSecret, err := sdk.ReadSecret("payload-secret")
+
+	if err != nil {
+		return fmt.Errorf("couldn't get payload-secret: %t", err)
+	}
+
+	xCloudSignature := r.Header.Get(sdk.CloudSignatureHeader)
+
+	err = hmac.Validate(*req, xCloudSignature, payloadSecret)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
